refactor: 追加MODのリストを別ファイルに分離
